distributed devs to src
