+++ test/DistributedTests/runtests.jl
@@ -1,0 +1,5 @@
+module DistributedTests
+
+include("sequential/runtests.jl")
+
+end # module
