+++ test/DistributedTests/sequential/PoissonTests.jl
@@ -1,0 +1,7 @@
+module PoissonTestsSeq
+using PartitionedArrays
+include("../PoissonTests.jl")
+with_debug() do distribute
+    PoissonTests.main(distribute,(2,2))
+end
+end
