+++ test/DistributedTests/sequential/runtests.jl
@@ -1,0 +1,7 @@
+module SequentialTests
+
+using Test
+
+@time @testset "PoissonSeq" begin include("PoissonTests.jl") end
+
+end
